* upstream/master: (671 commits)
  Bug 1103623 follow-up to fuzz 633344-1.html for D2D on Windows 8
  Bumping manifests a=b2g-bump
  Bumping gaia.json for 2 gaia revision(s) a=gaia-bump
  Bumping manifests a=b2g-bump
  Bumping gaia.json for 2 gaia revision(s) a=gaia-bump
  Bumping manifests a=b2g-bump
  Bumping gaia.json for 2 gaia revision(s) a=gaia-bump
  Bumping manifests a=b2g-bump
  Bumping gaia.json for 2 gaia revision(s) a=gaia-bump
  Bug 1103071 - Modify test_switch_remote_frame for compatibility with e10s. r=jgriffin
  Bug 996397 - B2G NFC: Replace DOMRequests with Promises. r=smaug, allstars.chh
  Bumping manifests a=b2g-bump
  Bumping gaia.json for 4 gaia revision(s) a=gaia-bump
  Bug 1094572 - B2G RIL: handle data call fail cause accordingly. r=echen
  Bug 1072922 - Add libc constants for flock structure type. r=Yoric
  Bug 1095521 -  Implement more paste commands for the inspector panel. r=pbrosset
  Bug 1074106 - [timeline] marker sidebar. r=vporof
  Bug 1103640 followup - fix syntax error, r=me
  Bug 1103640 - Allow DeviceActor to work when no browser windows. r=jryans
  Bug 1017931 - allow about:accounts to be used even if legacy sync is configured. r=adw
  ...

Conflicts:
	browser/components/loop/content/shared/js/activeRoomStore.js
	browser/components/loop/standalone/content/js/standaloneMozLoop.js
	browser/components/loop/test/shared/activeRoomStore_test.js
	browser/components/loop/test/standalone/standaloneMozLoop_test.js
	browser/devtools/framework/toolbox-options.js
	browser/devtools/jar.mn
	browser/devtools/performance/controller.js
	browser/devtools/performance/performance.xul
	browser/devtools/performance/test/browser.ini

+++ browser/components/loop/content/js/conversation.js
@@ -20,6 +20,9 @@ loop.conversation = (function(mozL10n) {
   var CallIdentifierView = loop.conversationViews.CallIdentifierView;
   var DesktopRoomConversationView = loop.roomViews.DesktopRoomConversationView;
 
+  // Matches strings of the form "<nonspaces>@<nonspaces>" or "+<digits>"
+  var EMAIL_OR_PHONE_RE = /^(:?\S+@\S+|\+\d+)$/;
+
   var IncomingCallView = React.createClass({displayName: 'IncomingCallView',
     mixins: [sharedMixins.DropdownMenuMixin, sharedMixins.AudioMixin],
 
@@ -298,13 +301,13 @@ loop.conversation = (function(mozL10n) {
 
           document.title = mozL10n.get("conversation_has_ended");
 
-          var feebackAPIBaseUrl = navigator.mozLoop.getLoopCharPref(
+          var feebackAPIBaseUrl = navigator.mozLoop.getLoopPref(
             "feedback.baseUrl");
 
           var appVersionInfo = navigator.mozLoop.appVersionInfo;
 
           var feedbackClient = new loop.FeedbackAPIClient(feebackAPIBaseUrl, {
-            product: navigator.mozLoop.getLoopCharPref("feedback.product"),
+            product: navigator.mozLoop.getLoopPref("feedback.product"),
             platform: appVersionInfo.OS,
             channel: appVersionInfo.channel,
             version: appVersionInfo.version
@@ -505,14 +508,27 @@ loop.conversation = (function(mozL10n) {
     declineAndBlock: function() {
       navigator.mozLoop.stopAlerting();
       var token = this.props.conversation.get("callToken");
-      this.props.client.deleteCallUrl(token,
-        this.props.conversation.get("sessionType"),
-        function(error) {
+      var callerId = this.props.conversation.get("callerId");
+
+      // If this is a direct call, we'll need to block the caller directly.
+      if (callerId && EMAIL_OR_PHONE_RE.test(callerId)) {
+        navigator.mozLoop.calls.blockDirectCaller(callerId, function(err) {
           // XXX The conversation window will be closed when this cb is triggered
           // figure out if there is a better way to report the error to the user
-          // (bug 1048909).
-          console.log(error);
+          // (bug 1103150).
+          console.log(err.fileName + ":" + err.lineNumber + ": " + err.message);
         });
+      } else {
+        this.props.client.deleteCallUrl(token,
+          this.props.conversation.get("sessionType"),
+          function(error) {
+            // XXX The conversation window will be closed when this cb is triggered
+            // figure out if there is a better way to report the error to the user
+            // (bug 1048909).
+            console.log(error);
+          });
+      }
+
       this._declineCall();
     },
 
@@ -616,10 +632,10 @@ loop.conversation = (function(mozL10n) {
     // don't work in the conversation window
     window.OT.overrideGuidStorage({
       get: function(callback) {
-        callback(null, navigator.mozLoop.getLoopCharPref("ot.guid"));
+        callback(null, navigator.mozLoop.getLoopPref("ot.guid"));
       },
       set: function(guid, callback) {
-        navigator.mozLoop.setLoopCharPref("ot.guid", guid);
+        navigator.mozLoop.setLoopPref("ot.guid", guid);
         callback(null);
       }
     });
@@ -641,13 +657,11 @@ loop.conversation = (function(mozL10n) {
       dispatcher: dispatcher,
       sdkDriver: sdkDriver
     });
-    var activeRoomStore = new loop.store.ActiveRoomStore({
-      dispatcher: dispatcher,
+    var activeRoomStore = new loop.store.ActiveRoomStore(dispatcher, {
       mozLoop: navigator.mozLoop,
       sdkDriver: sdkDriver
     });
-    var roomStore = new loop.store.RoomStore({
-      dispatcher: dispatcher,
+    var roomStore = new loop.store.RoomStore(dispatcher, {
       mozLoop: navigator.mozLoop,
       activeRoomStore: activeRoomStore
     });
